Update config file and refactor repo structure
