+++ client_src/internal/domain/config/server.go
@@ -2,6 +2,7 @@ package config
 
 type Config struct {
 	Servers []ServerConfig `yaml:"servers"`
+	Watcher WatcherConfig  `yaml:"watcher,omitempty"`
 }
 
 type ServerConfig struct {
@@ -12,4 +13,13 @@ type ServerConfig struct {
 type UserConfig struct {
 	Username string `yaml:"username"`
 	Password string `yaml:"password"`
+}
+
+type WatcherConfig struct {
+	Reconnect ReconnectConfig `yaml:"reconnect,omitempty"`
+}
+
+type ReconnectConfig struct {
+	Timeout int `yaml:"timeout"`
+	Retries int `yaml:"retries"`
 }
