+++ client_src/internal/domain/connections/consts.go
