+++ client_src/internal/domain/connections/interface.go
