+++ client_src/internal/domain/template/consts.go
