+++ client_src/internal/parsers/files.go
