+++ client_src/internal/service/adapters/linux/config_filler.go
@@ -4,9 +4,9 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
-	"implementation/internal/domain/config"
-	"implementation/internal/domain/template"
-	"implementation/internal/parsers"
+	"implementation/client_src/internal/domain/config"
+	"implementation/client_src/internal/domain/template"
+	"implementation/client_src/internal/parsers"
 	"log"
 	"os"
 	"path"
