+++ client_src/internal/service/connection_service.go
@@ -2,7 +2,7 @@ package service
 
 import (
 	"fmt"
-	"implementation/internal/domain/connections"
+	"implementation/client_src/internal/domain/connections"
 	"log"
 	"time"
 )
