+++ client_src/internal/service/iface_service.go
@@ -4,11 +4,11 @@ import (
 	"errors"
 	"fmt"
 	"github.com/mdlayher/ndp"
-	"implementation/internal/domain/config"
-	"implementation/internal/domain/connections"
-	"implementation/internal/domain/template"
-	"implementation/internal/parsers"
-	"implementation/internal/service/adapters/network"
+	"implementation/client_src/internal/domain/config"
+	"implementation/client_src/internal/domain/connections"
+	"implementation/client_src/internal/domain/template"
+	"implementation/client_src/internal/parsers"
+	"implementation/client_src/internal/service/adapters/network"
 	"log"
 	"net"
 	"net/netip"
